requirements.txt
# Flask & Extensions
flask==3.0.0
flask-wtf==1.2.1
python-dotenv==1.0.0
flask-sqlalchemy==3.1.1

# Data Processing
pandas>=2.2.0
openpyxl==3.1.2
numpy>=1.26.0

# Utils
requests==2.31.0
paramiko==3.3.1  # SSH 연결용
netmiko==4.2.0   # 네트워크 장비 연결용 